updated test entity, created CreateTestCommand.cs and dto for it

+++ src/Domain/Entities/TestEntity.cs
@@ -4,7 +4,7 @@ public class TestEntity : BaseEntity
 {
     public string? Title { get; set; }
     public string? Description { get; set; }
-    public double? MaxMark { get; set; }
+    public decimal? MaxMark { get; set; }
     public UserEntity? UserCreator { get; set; }
     public DateTime? CreatedDate { get; set; }
     public DateTime? LastModifiedDate { get; set; }
